refactor: first-date-label 컴포넌트 분리
