+++ src/shared/ui/calendar/calendar-cell-date.tsx
@@ -1,8 +1,9 @@
 import { cva, type VariantProps } from 'class-variance-authority'
 
-import type { ReactNode } from 'react'
-import Text from '../text/text'
 import { cn } from '../../utils'
+import Text from '../text/text'
+
+import type { ReactNode } from 'react'
 
 /**
  * @description 날짜 셀 컴포넌트에서 사용되는 버튼입니다.
