+++ src/shared/ui/calendar/calendar-header.tsx
@@ -1,7 +1,8 @@
 import { format } from 'date-fns'
 
-import { useCalendarContext } from './hooks/use-calendar-context'
 import Text from '../text/text'
+
+import { useCalendarContext } from './hooks/use-calendar-context'
 
 /**
  * @description  YYYY.MM 표시 헤더
