+++ src/shared/ui/calendar/calendar-month.tsx
@@ -1,5 +1,6 @@
-import CalendarCell from './calendar-cell'
 import { isSameDay, isSameMonth, isToday } from 'date-fns'
+
+import CalendarCell from './calendar-cell'
 import { useCalendarContext } from './hooks/use-calendar-context'
 
 /**
