+++ src/shared/ui/calendar/hooks/index.tsx
@@ -1,4 +1,3 @@
-export { default as useCalendarDates } from './use-calendar-date'
 export { useCalendarContext, CalendarProvider } from './use-calendar-context'
 export { default as useMonthObserver } from './use-month-observer'
 export { default as useAllMonthDates } from './use-all-month-dates'
