+++ src/shared/ui/calendar/index.tsx
@@ -4,6 +4,7 @@ export { default as WeekdaysHeader } from './weekday-header'
 export { default as CalendarMonth } from './calendar-month'
 export { default as CalendarDateCell } from './calendar-cell'
 export { default as CalendarCellDate } from './calendar-cell-date'
+export { default as CalendarFirstDateLabel } from './calendar-fitst-date-label'
 export { default as CalendarScrollWrapper } from './calendar-scroll-wrapper'
 export { default as ScheduleBadgeFill } from './schedule-badge-fill'
 export { default as ScheduleBadgeMarker } from './schedule-badge-marker'
