+++ src/shared/ui/calendar/schedule-badge-marker.tsx
@@ -1,4 +1,5 @@
 import { cva, type VariantProps } from 'class-variance-authority'
+
 import Text from '../text/text'
 
 /**
