+++ src/shared/ui/calendar/weekday-header.tsx
@@ -1,4 +1,5 @@
 import { cva } from 'class-variance-authority'
+
 import Text from '../text/text'
 
 /**
